Pump Control/requirements.txt
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1733250440834/work
Bottleneck @ file:///C:/b/abs_816hr2khp1/croot/bottleneck_1731058648110/work
Brotli @ file:///C:/b/abs_c415aux9ra/croot/brotli-split_1736182803933/work
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
comm @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_comm_1753453984/work
contourpy @ file:///C:/b/abs_f2u2o_8s9g/croot/contourpy_1732540071787/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
debugpy @ file:///D:/bld/bld/rattler-build_debugpy_1758162072/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1740384970518/work
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1746947292760/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1756729339227/work
fonttools @ file:///C:/miniconda3/conda-bld/fonttools_1758791708108/work
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_importlib-metadata_1747934053/work
ipykernel @ file:///D:/bld/ipykernel_1754352877088/work
ipython @ file:///D:/bld/bld/rattler-build_ipython_1759151854/work
ipython_pygments_lexers @ file:///home/conda/feedstock_root/build_artifacts/ipython_pygments_lexers_1737123620466/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1733440914442/work
jupyter_core @ file:///D:/bld/jupyter_core_1748333031907/work
kiwisolver @ file:///C:/b/abs_faf90xet7a/croot/kiwisolver_1737040915779/work
matplotlib==3.10.6
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1733416936468/work
mkl-service==2.4.0
mkl_fft @ file:///C:/b/abs_64riv3lfyn/croot/mkl_fft_1751310924052/work
mkl_random @ file:///C:/b/abs_47jvsmau4j/croot/mkl_random_1751310922820/work
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
numexpr @ file:///C:/b/abs_3buvfgcf6r/croot/numexpr_1755766488094/work
numpy @ file:///C:/miniconda3/conda-bld/numpy_and_numpy_base_1757670256083/work/dist/numpy-2.3.3-cp313-cp313-win_amd64.whl#sha256=d9474e66189cba52ea8e11316ddea0a38ae7f8fc06e3ef798d9bbfddc6ba51a5
packaging @ file:///C:/b/abs_f7t3vr2_lo/croot/packaging_1753775377650/work
pandas @ file:///C:/b/abs_58a1hlvjeo/croot/pandas_1756466410850/work/dist/pandas-2.3.2-cp313-cp313-win_amd64.whl#sha256=e3d25986bd6f7ec53acbba8bf94ee4f49c11b115835d0b9a4c62809e4f788a80
parso @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_parso_1755974222/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
pillow @ file:///C:/b/abs_b50vowcrzo/croot/pillow_1738010273782/work
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1756227402/work
ply @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/ply_1729041337918/work
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1756321756983/work
psutil @ file:///D:/bld/psutil_1758169202226/work
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1750615794071/work
pyparsing @ file:///C:/Users/dev-admin/buildout/croot/pyparsing_1735861809022/work
PyQt5==5.15.10
PyQt5_sip @ file:///C:/b/abs_d1i1jr9fgt/croot/pyqt-split_1736540551157/work/pyqt_sip
pyqtgraph @ file:///C:/b/abs_28oet7jt1w/croot/pyqtgraph_1744105277815/work
pyserial @ file:///C:/b/abs_38q410cihh/croot/pyserial_1736540571753/work
python-dateutil @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/python-dateutil_1729038406576/work
pytz @ file:///C:/b/abs_f8wdzeix0n/croot/pytz_1752135878094/work
pywin32 @ file:///D:/bld/bld/rattler-build_pywin32_1756487139/work
pyzmq @ file:///D:/bld/bld/rattler-build_pyzmq_1757387025/work
setuptools==72.1.0
sip @ file:///C:/b/abs_b6twfwtueg/croot/sip_1736544693649/work
six @ file:///C:/b/abs_149wuyuo1o/croot/six_1744271521515/work
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
tornado @ file:///D:/bld/tornado_1756854854126/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1756220668/work
tzdata @ file:///croot/python-tzdata_1746123641790/work
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1758622279606/work
wheel==0.45.1
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1749421620841/work
